moved the getting builder logic to api route file
